feat: adicionar ESLint e reorganizar estrutura do projeto
