--- commands/geral/oi.js
@@ -1,9 +1,0 @@
-module.exports = {
-	name: "oi",
-	description: "respondo oi",
-	type: 1,
-
-	run: async (client, interaction, args) => {
-		interaction.reply(`**oi ${interaction.user.username} 👋**`);
-	},
-};
