--- handlers/db.js
@@ -1,7 +1,0 @@
-const mongoose = require("mongoose");
-
-mongoose.connect(process.env.MONGODB_URI);
-
-let db = mongoose.connection;
-
-module.exports = db;
