--- handlers/events.js
@@ -1,19 +1,0 @@
-const { readdirSync } = require("fs");
-
-module.exports = (client) => {
-  const load = (dirs) => {
-    const events = readdirSync(`./events/${dirs}/`).filter((f) =>
-      f.endsWith(".js")
-    );
-    for (let file of events) {
-      const event = require(`../events/${dirs}/${file}`);
-      let eventName = file.split(".")[0];
-      console.log(`O evento ${eventName} foi carregado com sucesso!`);
-
-      client.on(eventName, event.bind(null, client));
-    }
-  };
-  readdirSync("./events/").forEach((dirs) => {
-      load(dirs);
-   });
-};
