+++ package.json
@@ -1,21 +1,40 @@
 {
-  "name": "ifchatbot",
-  "version": "0.0.1",
-  "main": "bot.js",
-  "scripts": {
-    "start": "node bot.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@discordjs/voice": "^0.14.0",
-    "discord-speech-to-text": "^1.1.5",
-    "discord.js": "^14.7.1",
-    "dotenv": "^16.0.3",
-    "libsodium-wrappers": "^0.7.10",
-    "mongoose": "^6.8.3",
-    "prism-media": "^2.0.0-alpha.0"
-  },
-  "description": ""
+	"name": "ifchatbot",
+	"version": "0.0.1",
+	"main": "bot.js",
+	"type": "commonjs",
+	"scripts": {
+		"start": "node ./src/bot.js",
+		"test": "echo \"Error: no test specified\" && exit 1",
+		"lint": "eslint 'src/**/*.{js,ts}' --fix"
+	},
+	"author": "",
+	"license": "ISC",
+	"dependencies": {
+		"@discordjs/opus": "^0.9.0",
+		"@discordjs/voice": "^0.14.0",
+		"discord-speech-to-text": "^1.1.5",
+		"discord.js": "^14.7.1",
+		"dotenv": "^16.0.3",
+		"libsodium-wrappers": "^0.7.10",
+		"mongoose": "^6.8.3",
+		"prism-media": "^2.0.0-alpha.0"
+	},
+	"description": "",
+	"devDependencies": {
+		"@types/mongoose": "^5.11.97",
+		"@types/node": "^18.14.6",
+		"@typescript-eslint/eslint-plugin": ">=5.43.0",
+		"@typescript-eslint/parser": ">=5.43.0",
+		"eslint": ">=8.0.0",
+		"eslint-config-prettier": "^8.7.0",
+		"eslint-plugin-node": "^11.1.0",
+		"eslint-plugin-prettier": "^4.2.1",
+		"prettier": "^2.8.4",
+		"ts-node": "^10.9.1",
+		"typescript": ">=4.4"
+	},
+	"engines": {
+		"node": ">=10.0.0"
+	}
 }
