+++ src/assets/textos/curiosidades.json
