+++ src/handlers/db.js
@@ -1,0 +1,7 @@
+const mongoose = require("mongoose");
+
+mongoose.connect(process.env.MONGODB_URI);
+
+const db = mongoose.connection;
+
+module.exports = db;
